put all makeARGB tests together; name for clarity
