+++ pyqtgraph/tests/test_makeARGB.py
@@ -1,9 +1,11 @@
+# -*- coding: utf-8 -*-
 import numpy as np
 import pytest
+import sys
 from typing import Dict, Any, Union, Type
 
 from pyqtgraph import getCupy, setConfigOption
-from pyqtgraph.tests.test_functions import makeARGB
+from pyqtgraph.functions import makeARGB as real_makeARGB
 
 IN_2D_INT8 = np.array([[173, 48, 122, 41], [210, 192, 0, 5], [104, 56, 102, 115], [78, 19, 255, 6]], dtype=np.uint8)
 IN_RGB_INT8 = np.array(
@@ -4239,6 +4241,17 @@ def _do_something_for_every_combo(func):
                             func(data, key, levels, lut, scale, use_rgba)
 
 
+def _makeARGB(*args, **kwds):
+    img, alpha = real_makeARGB(*args, **kwds)
+    if kwds.get('useRGBA'):  # endian independent
+        out = img
+    elif sys.byteorder == 'little':  # little-endian ARGB32 to B,G,R,A
+        out = img
+    else:  # big-endian ARGB32 to B,G,R,A
+        out = img[..., [3, 2, 1, 0]]
+    return out, alpha
+
+
 def save_reference():
     """
     This saves the output (or exception type) of running makeARGB for every combo of arguments. The
@@ -4249,7 +4262,7 @@ def save_reference():
 
         def write_expectation_to_file(data, key, levels, lut, scale, use_rgba):
             try:
-                output, alpha = makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
+                output, alpha = _makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
             except Exception as e:
                 tmp_file.write(f"{key!r}: {type(e)}\n")
             else:
@@ -4258,26 +4271,26 @@ def save_reference():
         _do_something_for_every_combo(write_expectation_to_file)
 
 
-def test_makeARGB():
+def test_makeARGB_against_generated_references():
     def assert_correct(data, key, levels, lut, scale, use_rgba):
         expectation = EXPECTED_OUTPUTS[key]
         if isinstance(expectation, type) and issubclass(expectation, Exception):
             try:
-                makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
+                _makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
             except Exception as e:
                 assert expectation == type(e)
             else:
                 assert False, f"makeARGB({key!r}) was supposed to raise {expectation} but didn't raise anything."
         else:
-            output, alpha = makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
+            output, alpha = _makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
             assert (
                 output == expectation
-            ).all(), f"Incorrect makeARGB({key!r}) output! Expected:\n{expectation!r}\n  Got:\n{output!r}"
+            ).all(), f"Incorrect _makeARGB({key!r}) output! Expected:\n{expectation!r}\n  Got:\n{output!r}"
 
     _do_something_for_every_combo(assert_correct)
 
 
-def test_cupy_makeARGB():
+def test_cupy_makeARGB_against_generated_references():
     setConfigOption("useCupy", True)
     cp = getCupy()
     if cp is None:
@@ -4290,16 +4303,203 @@ def test_cupy_makeARGB():
         expectation = EXPECTED_OUTPUTS[key]
         if isinstance(expectation, type) and issubclass(expectation, Exception):
             try:
-                makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
+                _makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
             except Exception as e:
                 assert expectation == type(e)
             else:
                 assert False, f"makeARGB({key!r}) was supposed to raise {expectation} but didn't raise anything."
         else:
             expectation = cp.asarray(expectation)
-            output, alpha = makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
+            output, alpha = _makeARGB(data, lut=lut, levels=levels, scale=scale, useRGBA=use_rgba)
             assert (
                 output == expectation
-            ).all(), f"Incorrect makeARGB({key!r}) output! Expected:\n{expectation!r}\n  Got:\n{output!r}"
+            ).all(), f"Incorrect _makeARGB({key!r}) output! Expected:\n{expectation!r}\n  Got:\n{output!r}"
 
     _do_something_for_every_combo(assert_cupy_correct)
+
+
+def test_makeARGB_with_human_readable_code():
+    # Many parameters to test here:
+    #  * data dtype (ubyte, uint16, float, others)
+    #  * data ndim (2 or 3)
+    #  * levels (None, 1D, or 2D)
+    #  * lut dtype
+    #  * lut size
+    #  * lut ndim (1 or 2)
+    #  * useRGBA argument
+    # Need to check that all input values map to the correct output values, especially
+    # at and beyond the edges of the level range.
+
+    def checkArrays(a, b):
+        # because py.test output is difficult to read for arrays
+        if not np.all(a == b):
+            comp = []
+            for i in range(a.shape[0]):
+                if a.shape[1] > 1:
+                    comp.append('[')
+                for j in range(a.shape[1]):
+                    m = a[i, j] == b[i, j]
+                    comp.append('%d,%d  %s %s  %s%s' %
+                                (i, j, str(a[i, j]).ljust(15), str(b[i, j]).ljust(15),
+                                 m, ' ********' if not np.all(m) else ''))
+                if a.shape[1] > 1:
+                    comp.append(']')
+            raise Exception("arrays do not match:\n%s" % '\n'.join(comp))
+
+    def checkImage(img, check, alpha, alphaCheck):
+        assert img.dtype == np.ubyte
+        assert alpha is alphaCheck
+        if alpha is False:
+            checkArrays(img[..., 3], 255)
+
+        if np.isscalar(check) or check.ndim == 3:
+            checkArrays(img[..., :3], check)
+        elif check.ndim == 2:
+            checkArrays(img[..., :3], check[..., np.newaxis])
+        elif check.ndim == 1:
+            checkArrays(img[..., :3], check[..., np.newaxis, np.newaxis])
+        else:
+            raise Exception('invalid check array ndim')
+
+    # uint8 data tests
+
+    im1 = np.arange(256).astype('ubyte').reshape(256, 1)
+    im2, alpha = _makeARGB(im1, levels=(0, 255))
+    checkImage(im2, im1, alpha, False)
+
+    im3, alpha = _makeARGB(im1, levels=(0.0, 255.0))
+    checkImage(im3, im1, alpha, False)
+
+    im4, alpha = _makeARGB(im1, levels=(255, 0))
+    checkImage(im4, 255 - im1, alpha, False)
+
+    im5, alpha = _makeARGB(np.concatenate([im1] * 3, axis=1), levels=[(0, 255), (0.0, 255.0), (255, 0)])
+    checkImage(im5, np.concatenate([im1, im1, 255 - im1], axis=1), alpha, False)
+
+    im2, alpha = _makeARGB(im1, levels=(128, 383))
+    checkImage(im2[:128], 0, alpha, False)
+    checkImage(im2[128:], im1[:128], alpha, False)
+
+    # uint8 data + uint8 LUT
+    lut = np.arange(256)[::-1].astype(np.uint8)
+    im2, alpha = _makeARGB(im1, lut=lut)
+    checkImage(im2, lut, alpha, False)
+
+    # lut larger than maxint
+    lut = np.arange(511).astype(np.uint8)
+    im2, alpha = _makeARGB(im1, lut=lut)
+    checkImage(im2, lut[::2], alpha, False)
+
+    # lut smaller than maxint
+    lut = np.arange(128).astype(np.uint8)
+    im2, alpha = _makeARGB(im1, lut=lut)
+    checkImage(im2, np.linspace(0, 127.5, 256, dtype='ubyte'), alpha, False)
+
+    # lut + levels
+    lut = np.arange(256)[::-1].astype(np.uint8)
+    im2, alpha = _makeARGB(im1, lut=lut, levels=[-128, 384])
+    checkImage(im2, np.linspace(191.5, 64.5, 256, dtype='ubyte'), alpha, False)
+
+    im2, alpha = _makeARGB(im1, lut=lut, levels=[64, 192])
+    checkImage(im2, np.clip(np.linspace(384.5, -127.5, 256), 0, 255).astype('ubyte'), alpha, False)
+
+    # uint8 data + uint16 LUT
+    lut = np.arange(4096)[::-1].astype(np.uint16) // 16
+    im2, alpha = _makeARGB(im1, lut=lut)
+    checkImage(im2, np.arange(256)[::-1].astype('ubyte'), alpha, False)
+
+    # uint8 data + float LUT
+    lut = np.linspace(10., 137., 256)
+    im2, alpha = _makeARGB(im1, lut=lut)
+    checkImage(im2, lut.astype('ubyte'), alpha, False)
+
+    # uint8 data + 2D LUT
+    lut = np.zeros((256, 3), dtype='ubyte')
+    lut[:, 0] = np.arange(256)
+    lut[:, 1] = np.arange(256)[::-1]
+    lut[:, 2] = 7
+    im2, alpha = _makeARGB(im1, lut=lut)
+    checkImage(im2, lut[:, None, ::-1], alpha, False)
+
+    # check useRGBA
+    im2, alpha = _makeARGB(im1, lut=lut, useRGBA=True)
+    checkImage(im2, lut[:, None, :], alpha, False)
+
+    # uint16 data tests
+    im1 = np.arange(0, 2 ** 16, 256).astype('uint16')[:, None]
+    im2, alpha = _makeARGB(im1, levels=(512, 2 ** 16))
+    checkImage(im2, np.clip(np.linspace(-2, 253, 256), 0, 255).astype('ubyte'), alpha, False)
+
+    lut = (np.arange(512, 2 ** 16)[::-1] // 256).astype('ubyte')
+    im2, alpha = _makeARGB(im1, lut=lut, levels=(512, 2 ** 16 - 256))
+    checkImage(im2, np.clip(np.linspace(257, 2, 256), 0, 255).astype('ubyte'), alpha, False)
+
+    lut = np.zeros(2 ** 16, dtype='ubyte')
+    lut[1000:1256] = np.arange(256)
+    lut[1256:] = 255
+    im1 = np.arange(1000, 1256).astype('uint16')[:, None]
+    im2, alpha = _makeARGB(im1, lut=lut)
+    checkImage(im2, np.arange(256).astype('ubyte'), alpha, False)
+
+    # float data tests
+    im1 = np.linspace(1.0, 17.0, 256)[:, None]
+    im2, alpha = _makeARGB(im1, levels=(5.0, 13.0))
+    checkImage(im2, np.clip(np.linspace(-128, 383, 256), 0, 255).astype('ubyte'), alpha, False)
+
+    lut = (np.arange(1280)[::-1] // 10).astype('ubyte')
+    im2, alpha = _makeARGB(im1, lut=lut, levels=(1, 17))
+    checkImage(im2, np.linspace(127.5, 0, 256).astype('ubyte'), alpha, False)
+
+    # nans in image
+
+    # 2d input image, one pixel is nan
+    im1 = np.ones((10, 12))
+    im1[3, 5] = np.nan
+    im2, alpha = _makeARGB(im1, levels=(0, 1))
+    assert alpha
+    assert im2[3, 5, 3] == 0  # nan pixel is transparent
+    assert im2[0, 0, 3] == 255  # doesn't affect other pixels
+
+    # 3d RGB input image, any color channel of a pixel is nan
+    im1 = np.ones((10, 12, 3))
+    im1[3, 5, 1] = np.nan
+    im2, alpha = _makeARGB(im1, levels=(0, 1))
+    assert alpha
+    assert im2[3, 5, 3] == 0  # nan pixel is transparent
+    assert im2[0, 0, 3] == 255  # doesn't affect other pixels
+
+    # 3d RGBA input image, any color channel of a pixel is nan
+    im1 = np.ones((10, 12, 4))
+    im1[3, 5, 1] = np.nan
+    im2, alpha = _makeARGB(im1, levels=(0, 1), useRGBA=True)
+    assert alpha
+    assert im2[3, 5, 3] == 0  # nan pixel is transparent
+
+    # test sanity checks
+    class AssertExc(object):
+        def __init__(self, exc=Exception):
+            self.exc = exc
+
+        def __enter__(self):
+            return self
+
+        def __exit__(self, *args):
+            assert args[0] is self.exc, "Should have raised %s (got %s)" % (self.exc, args[0])
+            return True
+
+    with AssertExc(TypeError):  # invalid image shape
+        pg.makeARGB(np.zeros((2,), dtype='float'))
+    with AssertExc(TypeError):  # invalid image shape
+        pg.makeARGB(np.zeros((2, 2, 7), dtype='float'))
+    with AssertExc():  # float images require levels arg
+        pg.makeARGB(np.zeros((2, 2), dtype='float'))
+    with AssertExc():  # bad levels arg
+        pg.makeARGB(np.zeros((2, 2), dtype='float'), levels=[1])
+    with AssertExc():  # bad levels arg
+        pg.makeARGB(np.zeros((2, 2), dtype='float'), levels=[1, 2, 3])
+    with AssertExc():  # can't mix 3-channel levels and LUT
+        pg.makeARGB(np.zeros((2, 2)), lut=np.zeros((10, 3), dtype='ubyte'), levels=[(0, 1)] * 3)
+    with AssertExc():  # multichannel levels must have same number of channels as image
+        pg.makeARGB(np.zeros((2, 2, 3), dtype='float'), levels=[(1, 2)] * 4)
+    with AssertExc():  # 3d levels not allowed
+        pg.makeARGB(np.zeros((2, 2, 3), dtype='float'), levels=np.zeros([3, 2, 2]))
